requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
bcrypt==4.1.2
bertopic==0.16.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cryptography==42.0.5
cymem==2.0.8
Cython==0.29.37
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastapi==0.110.0
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2024.2.0
gast==0.5.4
google-api-core==2.17.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
hdbscan==0.8.33
huggingface-hub==0.21.4
idna==3.6
install==1.3.5
Jinja2==3.1.3
joblib==1.3.2
jose==1.0.0
jwt==1.3.1
keras==2.15.0
langcodes==3.3.0
libclang==16.0.6
llvmlite==0.42.0
Markdown==3.5.2
MarkupSafe==2.1.5
ml-dtypes==0.2.0
motor==3.3.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
plotly==5.19.0
preshed==3.0.9
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
pymongo==4.6.2
pynndescent==0.5.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
sentry-sdk==1.40.6
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.38.2
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
umap-learn==0.5.5
urllib3==2.2.1
uvicorn==0.27.1
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.14.1
